code cleanup: removed uneccecary casts
[git-p4: depot-paths = "//skif/SkifKode/trunk/": change = 70497]

+++ skif-store-test/skif-store-test-api/src/main/java/no/statkart/skif/storetest/domain/AbstractStoreTestBubbleWithHistoryId.java
@@ -12,7 +12,7 @@ public abstract class AbstractStoreTestBubbleWithHistoryId<T extends AbstractSto
 
     @Override
     public Long getValue() {
-        return (Long)super.getValue();
+        return super.getValue();
     }
 
     public AbstractStoreTestBubbleWithHistoryId(Long value) {
